Revert "add cache.SetCleanFn"
This reverts commit c76b9d206bb64b9f5d3dea097c0ca5cf8c8a3887.

This is actually not valid -- the client can clean a value that is
currently in use by the end user. To do this properly, the client would
need to change to track the last access time of every value, and even
that isn't bulletproof (I can hold a value for a long time even though I
got it a while ago). Better is for the end user to not auto clean and
instead iterate over all values cleaning things as desired.