allow errors to expire immediately with -1 or 0
This brings the MaxErrAge behavior in line with MaxAge behavior.

Previously, if you wanted an error to expire immediately, you needed to
use a positive value for MaxErrAge. You could not use 0, because 0 would
opt into storing the value forever. You also could not use a negative
number, because negative ttl's were internally ignored. If you wanted
the error to not be cached, you had to use a tiny err expiry (i.e., 1).

Now, you can use negative or 0 for MaxErrAge to disable caching, same as
how MaxAge works.